Web now returns UUID's attached to Sockets and Steps
